+++ internal/model/entity/reminder_stock.go
@@ -3,7 +3,7 @@ package entity
 type ReminderStockEntity struct {
 	Code     string `json:"code" gorm:"size:50;unique;not null"`
 	Name     string `json:"name" gorm:"size:255;not null"`
-	MinStock int    `json:"min_stock" gorm:"default:0"`
+	MinStock int64  `json:"min_stock" gorm:"default:0"`
 }
 
 type ReminderStockModel struct {
